chore(ops): split ops.rs into focused submodules
ops.rs (2499 lines) → ops/{linear,norm,elementwise,embedding,sampling,attention,recurrent,tests}.rs.
Pure organizational refactor, no logic changes.

Also removes two spurious #[allow(dead_code)] annotations from recurrent.rs
and promotes tokenizer::decode to pub for example binary access.
