+++ src/ops/linear.rs
@@ -1,0 +1,140 @@
+use anyhow::Result;
+use cudarc::driver::{DevicePtr, DevicePtrMut};
+
+use crate::ffi;
+use crate::tensor::*;
+
+/// Matrix-vector multiplication: y = A @ x
+/// A: (M, K) row-major, x: (K,), y: (M,)
+pub fn gemv(ctx: &DeviceContext, a: &DeviceMatrix, x: &DeviceVec, y: &mut DeviceVec) -> Result<()> {
+    assert_eq!(a.cols, x.len, "A cols {} != x len {}", a.cols, x.len);
+    assert_eq!(a.rows, y.len, "A rows {} != y len {}", a.rows, y.len);
+
+    let (a_ptr, _ga) = a.data.device_ptr(&ctx.stream);
+    let (x_ptr, _gx) = x.data.device_ptr(&ctx.stream);
+    let (y_ptr, _gy) = y.data.device_ptr_mut(&ctx.stream);
+
+    unsafe {
+        ffi::gemv_cuda(
+            a_ptr as *const ffi::Half,
+            x_ptr as *const ffi::Half,
+            y_ptr as *mut ffi::Half,
+            a.rows as i32,
+            a.cols as i32,
+            ctx.stream.cu_stream(),
+        );
+    }
+
+    Ok(())
+}
+
+/// Linear layer: y = weight @ x
+pub(crate) fn linear(
+    ctx: &DeviceContext,
+    x: &DeviceVec,
+    weight: &DeviceMatrix,
+) -> Result<DeviceVec> {
+    let mut y = DeviceVec::zeros(ctx, weight.rows)?;
+    gemv(ctx, weight, x, &mut y)?;
+    Ok(y)
+}
+
+/// Fully fused MLP into pre-allocated output buffer
+pub fn fused_mlp_into(
+    ctx: &DeviceContext,
+    x: &DeviceVec,
+    gate_proj: &DeviceMatrix,
+    up_proj: &DeviceMatrix,
+    down_proj: &DeviceMatrix,
+    act: &mut DeviceVec,
+    out: &mut DeviceVec,
+) -> Result<()> {
+    assert_eq!(gate_proj.cols, x.len, "gate_proj cols != x len");
+    assert_eq!(up_proj.cols, x.len, "up_proj cols != x len");
+    assert_eq!(
+        gate_proj.rows, up_proj.rows,
+        "gate and up must have same output dim"
+    );
+    assert_eq!(
+        down_proj.cols, gate_proj.rows,
+        "down_proj cols != intermediate_size"
+    );
+    assert_eq!(down_proj.rows, out.len, "down_proj rows != out len");
+    assert_eq!(act.len, gate_proj.rows, "act len != intermediate_size");
+
+    let hidden_size = x.len;
+    let intermediate_size = gate_proj.rows;
+
+    let (x_ptr, _gx) = x.data.device_ptr(&ctx.stream);
+    let (gate_ptr, _gg) = gate_proj.data.device_ptr(&ctx.stream);
+    let (up_ptr, _gu) = up_proj.data.device_ptr(&ctx.stream);
+    let (down_ptr, _gd) = down_proj.data.device_ptr(&ctx.stream);
+    let (act_ptr, _ga) = act.data.device_ptr_mut(&ctx.stream);
+    let (out_ptr, _go) = out.data.device_ptr_mut(&ctx.stream);
+
+    unsafe {
+        ffi::fused_mlp_cuda(
+            x_ptr as *const ffi::Half,
+            gate_ptr as *const ffi::Half,
+            up_ptr as *const ffi::Half,
+            down_ptr as *const ffi::Half,
+            act_ptr as *mut ffi::Half,
+            out_ptr as *mut ffi::Half,
+            hidden_size as i32,
+            intermediate_size as i32,
+            ctx.stream.cu_stream(),
+        );
+    }
+
+    Ok(())
+}
+
+/// GEMM: Y = weight @ X (batched linear projection)
+/// weight: [out_dim, in_dim] row-major, X: HiddenStates [in_dim, seq_len], Y: HiddenStates [out_dim, seq_len]
+pub fn gemm(ctx: &DeviceContext, weight: &DeviceMatrix, x: &HiddenStates) -> Result<HiddenStates> {
+    let mut out = HiddenStates::zeros(ctx, weight.rows, x.seq_len)?;
+    gemm_into(ctx, weight, x, &mut out)?;
+    Ok(out)
+}
+
+/// GEMM into pre-allocated output buffer (zero allocation).
+pub(crate) fn gemm_into(
+    ctx: &DeviceContext,
+    weight: &DeviceMatrix,
+    x: &HiddenStates,
+    out: &mut HiddenStates,
+) -> Result<()> {
+    assert_eq!(
+        weight.cols, x.hidden_dim,
+        "weight cols {} != hidden_dim {}",
+        weight.cols, x.hidden_dim
+    );
+    assert_eq!(
+        out.hidden_dim, weight.rows,
+        "out hidden_dim {} != weight rows {}",
+        out.hidden_dim, weight.rows
+    );
+    assert_eq!(
+        out.seq_len, x.seq_len,
+        "out seq_len {} != x seq_len {}",
+        out.seq_len, x.seq_len
+    );
+
+    let (w_ptr, _gw) = weight.data.device_ptr(&ctx.stream);
+    let (x_ptr, _gx) = x.data.device_ptr(&ctx.stream);
+    let (y_ptr, _gy) = out.data.device_ptr_mut(&ctx.stream);
+
+    unsafe {
+        ffi::gemm_cuda(
+            w_ptr as *const ffi::Half,
+            x_ptr as *const ffi::Half,
+            y_ptr as *mut ffi::Half,
+            weight.rows as i32,
+            x.seq_len as i32,
+            weight.cols as i32,
+            ctx.stream.cu_stream(),
+        );
+    }
+
+    Ok(())
+}
