+++ src/ops/sampling.rs
@@ -1,0 +1,87 @@
+use anyhow::{Result, anyhow};
+use cudarc::driver::{CudaSlice, DevicePtr, DevicePtrMut};
+
+use crate::ffi;
+use crate::tensor::*;
+
+/// Argmax — returns the index of the maximum element
+pub fn argmax(ctx: &DeviceContext, x: &DeviceVec) -> Result<u32> {
+    let mut out_gpu: CudaSlice<i32> = ctx
+        .stream
+        .alloc_zeros(1)
+        .map_err(|e| anyhow!("Alloc failed: {}", e))?;
+
+    {
+        let (x_ptr, _gx) = x.data.device_ptr(&ctx.stream);
+        let (out_ptr, _go) = out_gpu.device_ptr_mut(&ctx.stream);
+
+        unsafe {
+            ffi::argmax_cuda(
+                x_ptr as *const ffi::Half,
+                out_ptr as *mut i32,
+                x.len as i32,
+                ctx.stream.cu_stream(),
+            );
+        }
+    }
+
+    // Sync before reading result from GPU to CPU
+    ctx.sync()?;
+
+    let result = ctx
+        .stream
+        .clone_dtoh(&out_gpu)
+        .map_err(|e| anyhow!("D2H copy failed: {}", e))?;
+
+    Ok(result[0] as u32)
+}
+
+/// GPU sampling: temperature → softmax → top-k → top-p → multinomial.
+/// Runs entirely on GPU. Only the final token ID (4 bytes) is transferred D2H.
+pub fn gpu_sample(
+    ctx: &DeviceContext,
+    logits: &DeviceVec,
+    probs_scratch: &mut CudaSlice<f32>,
+    params: &crate::sampler::SamplingParams,
+    random_val: f32,
+) -> Result<u32> {
+    let mut out_gpu: CudaSlice<i32> = ctx
+        .stream
+        .alloc_zeros(1)
+        .map_err(|e| anyhow!("Alloc failed: {}", e))?;
+
+    {
+        let (l_ptr, _gl) = logits.data.device_ptr(&ctx.stream);
+        let (p_ptr, _gp) = probs_scratch.device_ptr_mut(&ctx.stream);
+        let (o_ptr, _go) = out_gpu.device_ptr_mut(&ctx.stream);
+
+        let inv_temperature = if params.temperature > 0.0 {
+            1.0 / params.temperature
+        } else {
+            1.0
+        };
+
+        unsafe {
+            ffi::gpu_sample_cuda(
+                l_ptr as *const ffi::Half,
+                p_ptr as *mut f32,
+                o_ptr as *mut i32,
+                logits.len as i32,
+                inv_temperature,
+                params.top_k,
+                params.top_p,
+                random_val,
+                ctx.stream.cu_stream(),
+            );
+        }
+    }
+
+    ctx.sync()?;
+
+    let result = ctx
+        .stream
+        .clone_dtoh(&out_gpu)
+        .map_err(|e| anyhow!("D2H copy failed: {}", e))?;
+
+    Ok(result[0] as u32)
+}
